fix(guard): refresh decision-diff fixture and ratchet after runtime dedup
Regenerate the source-bound decision corpus fixture for the touched
decision-pipeline files, drop the stale duplicate-function baseline
entry for the removed bounded-sed parser pair, move the executable
flag contract into its own module so command_rules stays within its
oversized-file ratchet, and bind the strict caret policy through a
module-level partial so the evidence builders keep their function-size
budgets.
